Level_32
修复史莱姆越窜越高的bug
调整史莱姆碰撞体大小
新增史莱姆击败角色后的x坐标随机偏移

+++ Assets/Scripts/base/LevelData.cs
@@ -17,7 +17,7 @@ public class LevelItem
     public int ProgressID;
     public int TimeID;
     public int TimeID_max;
-    public int CreateTime;
+    public float CreateTime;
     public int EnemyType;
     public int Ran0;
     public int Ran1;
